requirements.txt
flask
faiss-cpu
transformers
scikit-learn
numpy
beautifulsoup4
requests
sentence-transformers
